GUI Scaling Refactor [ISSUE WITH TEXT RENDERING / LABEL AUTOSIZING]
Refactored the GUI scaling logic. The GUI now maximises its size to make itself the same size as the game window. Scaling logic has been applied within the existing UI components' drawing functions to accomodate for this change.

The scaled up fonts display a cut off bottom section, and the autosizing for the labels seems to be a bit off.