Implemented Button Functionality + Fixed GUI autosizing issue
The button component is now implemented. It is essentially a label with background and border turned on, that highlights when the mouse hovers over it, and is intended to work with the click event to trigger some script.

There was an issue with the GUI not sizing correctly when switching resolution, which has been fixed by replacing the dislay_set_gui_maximise() function by the display_set_gui_size and manually setting the size to the window dimensions.